Rename Always casez_encoder_8X3.v to Always casez_priority encoder_8X3.v
Build a priority encoder for 8-bit inputs. Given an 8-bit vector, the output should report the first (least significant) bit in the vector that is 1. Report zero if the input vector has no bits that are high. For example, the input 8'b10010000 should output 3'd4, because bit[4] is first bit that is high.

There is also a similar casex that treats both x and z as don't-care. I don't see much purpose to using it over casez.

The digit ? is a synonym for z. so 2'bz0 is the same as 2'b?0